Add gcp to schemas
