--- src/shared/gcp_agent/config_connector/iam_service_account.go
@@ -1,17 +1,0 @@
-package config_connector
-
-import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-type IAMServiceAccountSpec struct {
-	Description *string `json:"description,omitempty"`
-	Disabled    *bool   `json:"disabled,omitempty"`
-	DisplayName *string `json:"displayName,omitempty"`
-	ResourceID  *string `json:"resourceID,omitempty"`
-}
-
-type IAMServiceAccount struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	Spec IAMServiceAccountSpec `json:"spec,omitempty"`
-}
