--- src/shared/gcp_agent/types.go
@@ -1,5 +1,0 @@
-package gcp_agent
-
-const maxGCPNameLength = 30
-const truncatedHashLength = 6
-const maxTruncatedLength = maxGCPNameLength - truncatedHashLength - 1 // add another char for the hyphen
