+++ src/shared/gcpagent/config_conector.go
@@ -1,8 +1,7 @@
-package gcp_agent
+package gcpagent
 
 import (
 	"context"
-	"crypto/sha256"
 	gcpk8s "github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/k8s"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/clients/generated/apis/iam/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/clients/generated/apis/k8s/v1alpha1"
@@ -213,34 +212,4 @@ func (a *Agent) deleteGSAToKSAPolicy(ctx context.Context, c client.Client, names
 	}
 
 	return nil
-}
-
-func (a *Agent) generateGSAToKSAPolicyName(ksaName string) string {
-	return fmt.Sprintf("otr-%s-gcp-identity", ksaName)
-}
-
-func (a *Agent) generateGSADisplayName(namespace string, accountName string) string {
-	return fmt.Sprintf("otr-%s-%s-%s", a.clusterName, namespace, accountName)
-}
-
-func (a *Agent) generateGSAName(namespace string, accountName string) string {
-	// TODO: the max length is quite short, we should consider a different approach
-	fullName := a.generateGSADisplayName(namespace, accountName)
-
-	var truncatedName string
-	if len(fullName) >= maxTruncatedLength {
-		truncatedName = fullName[:maxTruncatedLength]
-	} else {
-		truncatedName = fullName
-	}
-
-	hash := fmt.Sprintf("%x", sha256.Sum256([]byte(fullName)))
-	hash = hash[:truncatedHashLength]
-
-	return fmt.Sprintf("%s-%s", truncatedName, hash)
-}
-
-func (a *Agent) GetGSAFullName(namespace string, accountName string) string {
-	gsaName := a.generateGSAName(namespace, accountName)
-	return fmt.Sprintf("%s@%s.iam.gserviceaccount.com", gsaName, a.projectID)
 }
