+++ src/shared/otterizecloud/graphqlclient/schema.graphql
@@ -653,6 +653,7 @@ type Intent {
 	databaseResources: [DatabaseConfig!]
 	awsActions: [String!]
 	azureRoles: [String!]
+	gcpPermissions: [String!]
 	internet: InternetConfig
 	status: IntentStatus
 }
@@ -668,6 +669,7 @@ input IntentInput {
 	databaseResources: [DatabaseConfigInput!]
 	awsActions: [String!]
 	azureRoles: [String!]
+	gcpPermissions: [String!]
 	internet: InternetConfigInput
 	status: IntentStatusInput
 }
@@ -705,6 +707,7 @@ type IntentsOperatorConfiguration {
 	kafkaACLEnforcementEnabled: Boolean!
 	istioPolicyEnforcementEnabled: Boolean!
 	awsIAMPolicyEnforcementEnabled: Boolean!
+	gcpIAMPolicyEnforcementEnabled: Boolean!
 	azureIAMPolicyEnforcementEnabled: Boolean!
 	databaseEnforcementEnabled: Boolean!
 	protectedServicesEnabled: Boolean!
@@ -720,6 +723,7 @@ input IntentsOperatorConfigurationInput {
 	protectedServicesEnabled: Boolean
 	egressNetworkPolicyEnforcementEnabled: Boolean
 	awsIAMPolicyEnforcementEnabled: Boolean
+	gcpIAMPolicyEnforcementEnabled: Boolean
 	azureIAMPolicyEnforcementEnabled: Boolean
 	databaseEnforcementEnabled: Boolean
 }
